src/core/boda_justice/serializers/offence/serializers.py
from boda_justice.models.offence.models import Offence
from rest_framework import serializers


class OffencesSerializer(serializers.ModelSerializer):

    class Meta:
        model = Offence
        fields = ('fine', 'description')
